include suggestions

+++ docker-images/notebook/prepare.sh
@@ -8,15 +8,19 @@ cp --update -r -v /pre-home/. /home/jovyan
 if [ -z "$EXAMPLES_GIT_URL" ]; then
     export EXAMPLES_GIT_URL=https://github.com/pangeo-data/pangeo-example-notebooks
 fi
-git clone $EXAMPLES_GIT_URL examples
+if [ ! -d "examples" ]; then
+  git clone $EXAMPLES_GIT_URL examples
+fi
 cd examples
 git remote set-url origin $EXAMPLES_GIT_URL
 git fetch origin
 git reset --hard origin/master
 git merge --strategy-option=theirs origin/master
-echo "Files in this directory should be treated as read-only"  > DONT_SAVE_ANYTHING_HERE.md
+if [ ! -f DONT_SAVE_ANYTHING_HERE.md ]; then
+  echo "Files in this directory should be treated as read-only"  > DONT_SAVE_ANYTHING_HERE.md
+fi
 cd ..
-mkdir work
+mkdir -p work
 
 if [ -e "/opt/app/environment.yml" ]; then
     echo "environment.yml found. Installing packages"
